Explicitly type state initialization in class components
Summary:
While I was testing something in relation to reviewing D436, I found out that Flow was picking up the type of `State` wrong. I investigated and realized it was occurring across the whole codebase.

Turns out that assigning the initial value of `state` without a type can confuse Flow. In particular, it gets confused about empty collections, and ends up typing them as "unsealed", which means that we get an `any` type when we iterate through those collections.

Luckily, it's easy to fix. We just need to explicitly declare the type of `this.state`.

Test Plan:
1. Make sure Flow reports no errors
2. In `AddUsersModal`, introduce some code that iterates through `userInfoInputArray` but has the wrong types. Confirm it didn't cause errors before because collection items were being `any`-typed, but now does
3. Use `flow type-at-pos` to examine types and make sure they are correct

Reviewers: KatPo, palys-swm

Reviewed By: KatPo

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D437