Use chatListDataWithNestedSidebars on web
Summary:
Currently, `chatListData` includes sidebars in the collection it returns, whereas `chatListDataWithNestedSidebars` doesn't. (Both of them include sidebars for each thread under the `sidebars` property of `ChatThreadItem`.)

This diff gets rid of the distinction, so we're back to just having a single `chatListData`, but now it doesn't include sidebars as top-level entities in the collection.

Test Plan: Flow

Reviewers: KatPo, palys-swm

Reviewed By: KatPo

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D442