[web] Merge keys in UPDATE_NAV_INFO
Summary:
Instead of overwriting the old `NavInfo`, now we will merge the old one with the new one.

This lets us avoid having to bind `navInfo` at every place we call `UPDATE_NAV_INFO`.

The only place we weren't previously merging the keys was on line 149 of `web/app.react.js`. To make sure this diff wouldn't change that behavior, I confirmed that the value that is set there – which comes from `navInfoFromURL` – includes values for all keys of `NavInfo`.

Test Plan: Flow

Reviewers: KatPo, palys-swm

Reviewed By: KatPo, palys-swm

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D458