[web] Always show active sidebar in ChatThreadList
Summary:
On `web`, unlike on `native`, the `ChatThreadList` is on-screen at the same time as `MessageList`. When showing a thread in the `MessageList`, we try to make sure that that thread is listed in the `ChatThreadList` too.

Since it will be possible to select a sidebar that isn't listed in the `ChatThreadList` through the `SidebarListModal`, we need to force them to be visible in a similar way.

Test Plan:
1. Construct a link to a valid sidebar that isn't shown by default. Load it and make sure the sidebar appears in the right place chronologically.
2. Make sure active sidebars aren't displayed as a separate thread at the top of the `ChatThreadList` anymore

Reviewers: KatPo, palys-swm

Reviewed By: KatPo, palys-swm

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D463