[lib] Don't require type parameter in mediaUpdate
Summary:
This diff fixes a bug on web where if you select an image for upload and immediately send the message, the message does not successfully send.

The bug was introduced in 46403debd4cd7d64552483f77d7956c3fadda65e. That diff expects `mediaUpdate` to always have a `type` parameter.

But we allow `$Shape<Media>` in `UpdateMultimediaMessageMediaPayload`, which means that a `type` parameter does not need to be set. We call it without `type` on web [here](https://github.com/Ashoat/squadcal/blob/master/web/input/input-state-container.react.js#L587-L589). That's what causes the bug.

Turns out we don't need the `invariant` to avoid the Flow error. In fact I think Flow is missing a type error, but I don't think the work necessary to fix the types here is really worth it.

Test Plan: Make sure the bug is fixed

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D489