[server] Warn when failing to fetch object during UpdateInfo hydration
Summary:
I had to test `create-personal-threads.js` multiple times today, and in between runs I would try to reset the database state with some queries. In my first attempt I forgot to clear out the `updates` table, and an `invariant` got triggered because the entry in the `threads` table had been cleared.

I shouldn't have used an `invariant` here. `invariant`s should only ever trigger if there is an error in the code, but in this case, database corruption was triggering the `invariant`.

This diff makes it so such database corruption no longer crashes the hydration process (`updateInfosFromRawUpdateInfos`), and adds a log statement that explains what happened.

Test Plan: Flow

Reviewers: palys-swm, KatPo

Reviewed By: KatPo

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D502