[native] Keep ChatThreadList sorted by time when search isn't active
Summary:
Had some confusion in D469. Didn't realize that it was changing the non-search behavior.

This diff also cleans up some of the logic so it's clearly gated on whether `searchText` is set. I think this is more readable, since it will be more clear to the reader that `usersSearchResults` won't be added to `chatItems` unless `searchText` is set.

Test Plan: Make sure the threads appear in the correct order again

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D507