[server] Add space between WHERE and ORDER BY statements in fetchServerThreadInfos
Summary:
While testing `create-personal-threads.js`, I accidentally passed in a `whereClause` that ended with an integer `threadID` instead of a string `threadID`.

This yielded a query that looked like `255ORDER BY` instead of `"255"ORDER BY`, which caused a SQL error. I fixed the type in my testing code, but this diff addresses: the real issue: the missing space.

Test Plan: Make sure that my testing code stopped yielding a SQL error, even though I was still passing in an integer instead of a string

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D509