[lib] Add getResultInfo to FetchJSONOptions
Summary:
This diff adds a new option to `fetchJSON` called `getResultInfo`. `getResultInfo` is a callback function that is guaranteed to be called before `fetchJSON` successfully resolves.

The callback is passed an object containing information about how the request was resolved. Right now there is just one parameter included: `interface`, which tells the caller whether the socket or the REST APIs were used to resolve a request.

The reason we need this will be explained in the next diff :)

Test Plan: Tested in combination with the next diff

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D510