[lib] Don't update message checkpoint on successful REST API message send
Summary:
When we're connected to a socket, we can increment the checkpoint on a successful message send because we know the socket is keeping us updated. Incrementing the checkpoint saves us the unnecessary work of fetching the message we already have locally the next time a socket connection is established.

However, incrementing the checkpoint on a successful REST API message send can lead to missing messages, since the checkpoint may pass over some messages that we haven't received yet, and the next time the socket connection is established, those messages won't be delivered.

This diff makes sure we only increment the checkpoint for socket sends.

Test Plan: I tested both socket and REST API message sends while looking at Redux through the React Native debugger and made sure the action payloads were correct

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D511