[native] Add "CONNECTING..." status to DisconnectedBar
Summary:
We'll show this new status when `showDisconnectedBar` is `false`, but the socket is still not connected.

The intention of this diff is that the user should always be aware if the socket is disconnected. But we don't want to display the scary red "DISCONNECTED" bar if it's just the app starting up.

Test Plan:
1. Turn the server on/off and make sure the red "DISCONNECTED" bar appears and reappears, without the gray "CONNECTING..." bar appearing
2. Background/foreground the app and make sure the gray "CONNECTING..." bar appears and reappears, without the red "DISCONNECTED" bar appearing
3. Test both dark mode and light mode

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D512