[lib] Do not include parent thread members in threadSearchText
Summary: When we changed threadInfo.members to include members of ancestor threads, we didn't change threadSearchText. Since we didn't switch to codeVersion 70 yet, this wasn't noticed, but when working on D516 results were different from native, and I discovered this was the cause

Test Plan:
Created thread with three users: A, B, C, created subthread with only user A, searched for user B:  no subthread in results.
Changed code version to 70, and searched for user B: subthread was in results.
Added role check to threadSearchText: subthread was not in results
Tested also with D516, compared web and native results for sidebars search, made sure the same sidebars were in results

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D517