[native] Display search component in pending thread message list
Summary: Introduce new nav param which tells if we should display a search component. This diff will be refactored in one of the next. The styling is not yet implemented.

Test Plan: When the param is not set, everthing should work as previously. When the param is set, user search component is displayed. Changing input value results in displaying real thread or a pending one.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D522