[native] Optimise thread search
Summary: Searching for a thread was taking acceptable amount of time, but the recent changes degraded the performance (allowing searching for all the threads, not only personal and searching for threads with multiple users).

Test Plan: Search for a thread: verify that the results are correct and that the performance significantly improved.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D523