Do not display blocked parent thread members and do not allow to add to subthread
Summary: When working on D528 I discovered that we're not handling blocked parent thread members correctly: they were listed without any notice and it was possible to select them without getting an alert. Server was handling this correctly and did not allow to add those users.

Test Plan: Checked if blocked parent thread members are displayed only as search results with correct notice, and selecting them displays an alert

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D531