[native] Include username in setNativeCredentials call in EditPassword
Summary:
While replacing all of the `$Shape`s in the codebase, I realized that there was a type error in `setNativeCredentials`, where it could sometimes fail to fetch the username and end up setting it to a falsey value.

This diff makes sure we always include the username we have from Redux.

Test Plan: Flow

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D556