[native] Require full credentials in setNativeCredentials
Summary:
We had code to fetch them if they were missing, but the typesystem actually missed an error there, because the fetch could possibly fail, in which case we could end up missing either the `username` or `password` (or both).

This diff fixes it by requiring both `username` and `password` to be set when passed into `setNativeCredentials`. D555 made sure all callsites conform to this new requirement.

Test Plan: Flow

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D557