[server] Better error logging for database backup logic
Summary:
I've been investigating an issue where database backups of the `squadcal` MySQL database keep on taking up all of the space of the backup directory, leaving no space for backups of the `phabricator` MySQL database. If I go in and delete the last several `squadcal` backups, things work fine for a while, but after I check a week later they've once again taken up all the space.

I'm still not sure why it's happening, but I noticed a couple corrupt `squadcal` backups in the directory, which was a bit strange because I expected them to get deleted if something failed during write.

It's possible that the corrupt backups are contributing to a pattern of disk utilization where the `squadcal` backups eat up 100% of remaining space, so that the next time the `phabricator` backup runs, it needs to delete two previous `phabricator` backups to have space for a new one (assuming the new one is slightly larger than the older ones).

I'm hoping this improved logging helps me understand why the corrupt `squadcal` backups appear, and if it might be connected to why they keep taking up all the space.

Test Plan: I carefully tested this on prod (since backups are disabled on dev)

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D561