[client] Show username in MessagePreview except if it matches thread name
Summary:
This diff makes us more picky about the scenarios in which we elide the username. Now we'll only do it for messages sent by the other user in a two-person thread that doesn't have a name (eg. default to the other user's username).

This change is in response to some feedback from a user that it was confusing for us to skip the username when the message was sent by the viewer.

Test Plan: Test with two clients, have one change the thread name to make sure that forces the username to appear

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D570