[server] Added private thread creator function
Summary: Added private-thread-creator function. This function currently does not set the squadbots role to -1, a future commit will do so.

Test Plan: Tested this individually and in combination with the full code. Individually creating threads still worked normally since this function isn't used yet. In combination with the other code squadbot was added with a role of -1 that I checked with phpMyAdmin, permissions were set to null, and also in the thread options squadbot was not listed as one of the members.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D648