[server] Send squadbot message at start of private thread
Summary: When a user creates an account a private thread will automatically be created. This private thread will have a message from squadbot introducing them to the private thread. Squadbot has a role of -1 and permissions of null, so the user will not be able to see squadbot in their members list on the thread settings.

Test Plan: Tested by creating a new account and seeing if the thread and the message was created. Checked the thread settings to make sure it did not say squadbot was a member. Made sure messages can be sent to the private thread and normal thread. Made sure groupchats still work, checked old private threads to make sure they still work. And also checked phpMyAdmin to make sure that squadbot was being created with the -1 role and with permissions set to NULL.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D650