remove viewer - every title contains author's name
Summary: having 2 users: (A,B), if user A sends a photo and creates a sidebar from it, the title is going to be: 'you sent a photo'. The problem is that the user B will see the same title: 'you sent a photo', although they haven't sent anything. The solution is to always put a user name in the thread title, no matter whether they created the message/sidebar or someone else.

Test Plan: Send a photo, create a sidebar from it and then see the title from the author's account and from some other user's account. Both users should see the author's name in the title.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat, palys-swm

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D707