[native] Fix allowImagePasteForThreadID @property annotation
Summary: The `@property` annotation for `allowImagePasteForThreadID` was incorrect. It happened to work previously but now it has been fixed to allow pending `threadInfo.id` to correctly be passed to the objc side.

Test Plan: Tested by pasting images into pending threads and seeing if the `ImagePasteModal` appeared. Next step is making adjustments to the `imagePastedCallback` to correctly handle sending of messages.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, subnub, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D734