[native] Use threadInfo and sourceMessageID in message list
Summary: In D727 I introduced a bug which caused the app to crash when e.g. opening thread settings. `getThreadIDFromParams` is used in a lot of different places and has to know about the nav params structure. There were two ways to fic the bug: change nav params shape in all the places or use the old approach in message list (we're still using optimistic thread info in camera modal, thread creator hook, etc.). Using optimistic thread in all the places doesn't make a lot of sense now - we would need to modify thread settings, thread delete and compose handlers, etc.

Test Plan: Open setting of a thread - the app should not crash. Send an image and check if it was send correctly.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D736