[server] Script to fix issues with previous new thread type scripts
Summary:
I noticed several issues with the way we migrated threads to become private threads recently, and one of them applied to the process we used for personal threads as well.

In general we should avoid custom logic going forward and try to use standard functions for changing complex things like thread type. It's also important we think really closely about the SQL queries we use in migration scripts.

Test Plan: Ran this on my local instance and confirmed it fixed things up

Reviewers: palys-swm, subnub

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D790