[lib] Break ties in message timestamps with message IDs
Summary: I ran a script that generated multiple messages with the same timestamp. `message-reducer` doesn't know how to break these ties, but when it uses the lower message ID, `activityUpdater` rejects it because it sees a newer message. We need to make sure `message-reducer` breaks ties in the same way that `activityUpdater` does.

Test Plan: loaded website and make sure it didn't crash

Reviewers: atul, palys-swm

Subscribers: KatPo, zrebcu411, Adrian, subnub, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D805