[web] Use hook instead of connect function in ResetPasswordModal
Test Plan: Added a console.log to the sendPasswordResetEmail server function so I could get the reset link. Reset the password with the link and checked if the password actually changed or not. Although, I do believe there are a couple issues that were also evident before this diff that I will explain in a comment

Reviewers: palys-swm, ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, KatPo, zrebcu411, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D849