[lib] Increase logIn timeout to 20s
Summary: The default for any HTTP request is a 10s timeout, but this isn't enough time for logging into my account on the Android emulator in release mode talking to prod servers. My account has a ton of data and thus a huge `log_in` payload, so this probably doesn't apply to other users. But still want to make sure I can log in on the emulator going forward.

Test Plan: Ran Android release build in emulator and made sure I could log in to the production server

Reviewers: atul, palys-swm

Reviewed By: atul

Subscribers: KatPo, zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D853