[lib] Don't trigger invariant in threadHasPermission for falsey threadInfo
Summary: This was broken in D389. The `invariant` is supposed to trigger only when a `RawThreadInfo` is passed in, but if a falsey `ThreadInfo` is passed in the function should return `false`. Right now, if a falsey `ThreadInfo` is passed in, the code decides it is a `RawThreadInfo`.

Test Plan: Make sure the invariant doesn't trigger in D856

Reviewers: atul, KatPo, palys-swm

Subscribers: zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D857