[lib] Trim newline characters from sidebar titles
Summary:
There was a bug in D693. For any text messages with a newline character, `getMessageTitle` would return the string with the newline at the end.

This meant that it was possible for threads to have titles that end with a newline character, which broke the sidebar's `MessageListHeaderTitle`.

We really should've caught this in the test plan... `getFirstNonQuotedRawLine` is supposed to just get the first line, not including any newline characters.

Test Plan: Create a new sidebar from a text message with multiple lines

Reviewers: atul, karol-bisztyga, palys-swm

Subscribers: KatPo, zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D860