[native] Avoid showing multiple lines of threadInfo.uiName in UI
Summary:
In D860, we were accidentally setting thread titles that ended with a newline character. I discovered that this could case issues in the UI.

That diff fixes `createPendingSidebar` so it doesn't include trailing newlines, but this diff tries to make it so the UI looks good enough for thread titles with multiple lines (just in case).

We had 10 places in native that rendered `uiName` in a single-line context. 7 of them were already using `SingleLine` or `firstLine`, but 3 of them weren't.

Test Plan: Make sure thread titles with multiple lines don't look weird in screen headers

Reviewers: atul, karol-bisztyga, palys-swm

Subscribers: KatPo, zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D861