[lib] Make sure bottom-right corner of pending sidebar message is rounded
Summary:
The danger of manually generating `ChatMessageInfoItem` (as we do in `getSourceMessageChatItemForPendingSidebar`) is that it becomes the developer's responsibility to set `startsConversation`, `startsCluster`, and `endsCluster` correctly.

This was done incorrectly in D579, leading to the pending sidebar message not having its bottom-right corner be rounded. This should've been caught during testing & review of D579. Going forward I guess I'll need to start asking for screenshots of every UI diff.

Test Plan: Make sure the corners are correctly rounded

Reviewers: atul, KatPo, palys-swm

Subscribers: zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D866