[lib] Simplify useMessageListData interface
Summary:
`boundMessageListData` can be computed inside the hook, so putting it there will simplify the interface.
This is a prerequisite for determining clustering properties of pending sidebar message: we need to have all the message infos instead of `ChatMessageItem`s provided to the hook.

Test Plan:
Send a message to an existing thread. Send a message to pending sidebar. Search for a thread and send a message.
Do the same for both native and web.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D874