[native] Do not hide log out button when username is long
Summary: There were two issues: `numberOfLines` in nested Text components doesn't work correctly and username inside `flex-direction: row` was taking to much space.

Test Plan:
Log in as a user with long username and check if log out button is fully visible.
Log in as a user with short username and check if space between label and username hasn't changed.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D877