[native] lifecycleState to enum, update useIsAppForegrounded
Summary: Made the switch from `boolean` to `LifecycleState` enum for `state.lifecycleState`, updated `useIsAppForegrounded` so 'prior' uses of the `state.foreground` property continue to work as before.

Test Plan: Tested on iOS and android simulator and things work as expected. Will take a thorough look at all of the uses of `foreground`/`NativeAppState` with the next diff to ensure there are no regressions.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D873