Basic db manager structure
Summary: This is a basic structure of database managers. It is designed in a way that they can be easily added and switched. iOS only for now. This is a POC so please do not mind the syntax etc., what matters is the overall direction and the idea.

Test Plan: none

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D889