Turn on Hermes on Android.
Summary: Turn on Hermes on Android. Not that it is advised to first clean the project and only then build & run it.

Test Plan: Launch the comm app on Android. If you can see it in Chrome Devtools(or in Flipper) to be debugged then it uses Hermews. Alternatively, you can do something like this: `const foo = (a,b) => a+b; console.log(foo.toString())`. If you see something like `'function foo(a0, a1) { [bytecode] }'` and you cannot see the function body, then the Herme is working.

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D883