[web] Check edit thread permission in thread settings
Summary: In blocked threads or private threads some thread settings might be disabled, and this causes errors. This diff adds permision check and disables input when there is no permission

Test Plan: Open private thread settings and notice error when trying to edit them. Add permission check and notice disabled inputs. Do the same with blocked thread, and verify that the behaviour is the same as in native

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D892