Enable typechecking for action functions
Summary:
Right now action functions aren't getting typechecked because we don't have a way to type a "bind" function that sets the first parameter of an input function and returns a function that takes the rest of the parameters.

This diff changes how we write action functions to make them possible to typecheck. It revealed several type errors (which I fixed).

Test Plan: Flow

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D897