[web] Handle error when threadInfo is not set in thread settings modal
Summary: Follow-up to D892: it might happen that thread will get invalidated when user has thread-settings-modal opened for this thread - thread might be deleted or user might leave/be removed. In this diff I made sure when this happens, user will see modal with explanation instead of an error, similar to native when we show alert

Test Plan:
Tested with D892:
Before this diff - opened thread in app and settings for this thread on web, in app left thread and noticed error on web. In this diff did the same thing, saw modal with explanation and checked if it closes on close button.
Tested the same scenario for removing user and deleting thread, with the same result.

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D912