Get rid of the `cloneDeep` from `lodash`
Summary: Assuming that the `cloneDeep` function has been introduced in order to avoid memory leaks/inefficient memory usage, an investigation checking whether it still occurs without cloning has been performed. Apparently this is not an issue anymore. The main reason for removing this function is our will to switch to Hermes. Using `cloneDeep` with this engine, a significant performance drop can be observed, the problem is described [here](https://github.com/facebook/hermes/issues/471).

Test Plan: Test results are presented in [this](https://phabricator.ashoat.com/D884#21096) and two following comments. We also ran a test logging in to the comm app with an account that has a lot of data to be fetched and parsed(and, up to this point, clonned).

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D916